docs: Add brief documentation comments to each function
