+++ src/scripts/runTests.ts
@@ -5,6 +5,11 @@ import { stepResponse } from "../main";
 export const TEST_SUCCESS = "✅ - All tests passed";
 export const TEST_ERROR = "⚠️ - Error running tests";
 
+/**
+ * Run tests and return the result
+ * @param coverageDir - Directory to store coverage report
+ * @returns Test result as a stepResponse object
+ */
 export const getTest = async (coverageDir: string): Promise<stepResponse> => {
   startGroup("Running tests");
   let response: stepResponse | undefined;
